requirements.txt
streamlit
pandas
reportlab
num2words
openpyxl
python-dotenv
mysql-connector-python>=8.0.33
